requirements.txt
pandas>=2.0.0
python-telegram-bot>=20.0
pymorphy3>=2.0.3
python-dotenv>=0.19.0
openpyxl>=3.0.0  # Для работы с Excel-файлами в pandas